first sentence revision
The first sentence, as stated, is too broad for what we will actually cover. Need to refer to warranty or remove the sentence.

On a side note, this section will need much further revision come 1902 as we start to allow the Corporate Procured Device model. The support areas will be not be as clear cut as we originally designed them to be.