Update data-loss-prevention-policies.md
Using labels as conditions does not explain that Exchange and Teams locations are not currently supported.  Customers receive the following error:

"Protecting labeled content in email and teams messages isn't supported. Either remove the label below or turn off Exchange and Teams as a location."

This is because Exchange transport does not evaluate the label metadata during message submission and delivery.