Update customizing.md
Changed references to "exemptions" to "exceptions" which is the term we are using to refer to deviations from the MMD baseline.  We need to make sure we are consistent in using "exceptions" to refer to this topic, not customizations or exemptions.

Also added a sentence to allow for MMD to revise the way in which exceptions are met so that we can change that "way" as the service matures.