Update ediscovery-cases.md
@markjjo 
Can we add the note about deleting a case.  Support has been getting few cases on this where a customer tries to delete a case ; however it still have  a hold, even though the hold is off are turned off the case can't be deleted.   Sometimes customer wants to delete the case and but the hold must be deleted before the case can be deleted.