Update view-the-dlp-reports.md with filter limitations
We need to update the documentation around scenario where a customer may have such a large amount of DLP policies that they may encounter 500 error upon attempting to get DLP report in SCC UI. This is due to all policies being returned by default for report and when customer has a large amount of DLP policies being returned the content URL length is exceeded, causing the 500 error. Customer must use a filter in this case in which they can filter DLP policies by service or by querying a group of policies.