Update status state for managing jobs-ediscovery20.md
We have had a few escalations where customers would like to know why they see a status of "Partially Successful" when performing an advanced eDiscovery search on custodian location.  However, it only indicates description for "Partially Successful" = "The job was partially successful."  When it actually means The job did not find any unindexed data in any of the custodian locations and the job is partially successful.  Would like documentation to be updated to include that in order to avoid more escalations going forward.