Update request for shared mailbox requirements
The mentioned statement is not correct in regards to shared mailboxes:
"Users can open protected mails in a shared mailbox where the shared mailbox received a protected mail as part of a distribution group."

The actual requirement (tested in lab and others. for more info on others, please reach me via mail or teams), for this setting to work is for the delegate not only to have full access but also to be automapped and this can be checked via cmd-let Get-MailboxPermission -identity "<shared@contoso.com>" -ReadFromDomainController | FL DelegateListLink" or "(Get-MailboxPermission -identity "<shared@contoso.com>" -ReadFromDomainController).DelegateListLink |sort -unique" to avoid duplicate values

This is the same as the on premises AD. The attribute that tells AD that the user has access to the shared is "MsExchDelegateListLink". On the cloud its via the the "DelegateListLink" that user permissions over shared mailbox sync to Azure AD to later be read by AIP. I can repro in lab anytime

Any questions or doubts please reach me (Manuel Pereira) or PM Martin Sieber .