Call out ExO cmdlets when being used
This is a long article and a customer has asked for us to add the need to use Exchange Online Powershell alongside SCC Powershell in the examples where ExO cmdlets are being used. The general need for ExO PS is written at the top of the article, but it would help to call-out where ExO cmdlets are being used in the examples, e.g. examples that call $DG = Get-DistributionGroup. The confusion has been that SCC also contains/contained the same PS cmdlets as ExO, but that those did not work from the SCC side.