[DO NOT MERGE] SEO edits for Create a secure guest sharing environment
Edited description to include keywords "guest access" and "collaboration".

"External sharing reviews" is a high-ranking keyword for this article, but I don't know if this is just another name for guest access reviews or something completely separate. Since there is nothing in the UI that specifically has "guest access reviews", I added it as an alternative name, but if this is incorrect I will remove it.