[DO NOT MERGE] SEO edits for Collaborate with guests in a team
Edited description to include keywords.

Edited body to include the keyword "document collaboration." I found a spot I felt it made sense to add. The article keeps the "collaboration" items pretty general, and no section is necessarily about tasks, conversation, or documentation; I added it to the file and folder sharing as the most likely place it fit.