[DO NOT MERGE] Edits for retention articles
Edited for the following items:

1. Changed either/or ordered lists to unordered lists.
2. Indented notes to align with associated steps where necessary.
3. Changed heading levels for consistency (jumped from 2 to 4 in some cases).
4. Removed colon at end of section in retention-policies-teams and replaced with period.

The author should verify the heading level for "When a user leaves the organization". In the Teams article, it is a stand-alone H2, but in Exchange and SharePoint, it's an H3 sub-heading. In both cases, the content doesn't seem directly associated with the H2 preceding it.