[DO NOT MERGE] SEO edits for security design principles
Edited metadata title to better reflect the wording of the trailhead article.

Added description and included keyword.

The SEO expert expressed that the metadata title at least should be changed. Google is ranking the article based on the "sail" in the title. I left the H1 alone since the subject of the article uses that express terminology, but edited the introduction to remove "sail."