Update back-up-data-before-switching-plans.md
Added a note to the document stating that "It's not required to back up data when you switch to another subscription that provides more and same services. User data will always be preserved and no action need be taken for any individual service that is present in both  expiring and replacement subscription assigned to the same user" as currently the users were under the assumption that it works but it needed clarity.