Correction on who can receive email notifications
Corrected a contradiction in the article. On line 60 it previously stated that "reviewers can be individual users, distribution or security groups, or Microsoft 365 groups". But further down the same page on line 89, there is a note saying that you can only specify users or mail-enabled security groups, and that Microsoft 365 groups are not supported. Through my own testing I found that the note is correct, e.g. you can only specify users or mail-enabled security groups. I changed line 60 in the article to remove the incorrect information (e.g. removed "Microsoft 365 groups" and changed "distribution or security groups" to say "mail-enabled security groups").