Update pilot-microsoft-365-from-my-custom-domain.md
This approach is best limited to email only and for a very small number of users for a pilot.  This article should specify those constraints so that users do not think that they can use this approach to start their migration to the cloud for all of the workloads and a large number of users.