Update configure-exchange-server-for-hybrid-modern-authentication.md
I believe the EvoSTS object created by the HCW at one time just named the object 'EvoSTS' so the powershell command to look it up may have worked at one point. That is no longer the case. The object is made with a name of 'EvoSts - GUID' so an 'eq' doesn't work anymore.

I changed this to 'like' with a wildcard on the end and then output to a table to show the full name and the enabled status.