Update convert-user-mailbox-to-shared-mailbox.md
Corrected "`Set-Mailbox -Identity mailbox1@contoso.onmicrosoft.com -Type Shared`" to "`Set-Mailbox -Identity mailbox1@contoso.com -Type Shared`"
In this example, the command is asked to be run in On-Premises. Customers should not use "<tenant>.onmicrosoft.com" domain in the On-Premises infrastructure, because this domains and changes related to this domain are not synced in Azure AD / MSODS. Is better to specify a vanity domain, instead of a Cloud default domain.