Encryption for external parties breaks under certain Conditional Access rules
I only found a vague reference to this in https://answers.microsoft.com/en-us/msoffice/forum/msoffice_o365admin-mso_exchon-mso_o365b/encrypted-e-mail-error-aadsts90072/e54062b1-1f6d-434d-bd1f-0f58b836de60 and as without this information and that policy in place this does not work and the error is not clear. Therefore added that CA policies need to be considered when sharing encrypted content to external users.