Update audit-log-retention-policies.md
We removed the "Create Audit Retention Policy" button on the main page.  So updated the verbiage in step 2 to have them click on the "Audit retention policies" tab.

Need to update the image in step 2 to remove the "Create Audit Retention Policy" button and to move the purple outline around the "Audit retention policies tab" instead.