added a note about converting an item from a list
For users that are not aware about reviewing list items, it takes a click of a button to convert a list object to a record.  From there these can then be reviewed in the excel file as well.

Let me know if this should be expanded out from a Tip to a walkthrough with pictures instead?