Updated SFTY section in Anti-Spam message headers
Removed legacy values that were initially intended to be implemented but never were implemented so these don't exist in the product today. We can add the info back if we decide to incorporate these in the future, but as the PM owner of impersonation & spoof, these are not in my planned list of items for the near term future. So I want to make sure the public doc is updated.