Update manage-addins-in-the-admin-center.md
Although the link for Services & Add-ins is correct and still points to the right place, the text no longer represents how the UI element is titled. Rather, it's now part of Settings > Org settings and Settings > Integrated apps based on which specific setting we're talking about.

Because of the new admin center, the experience to remove add-ins is also extremely convoluted. Instructions updated in this PR.

The note can also be removed because there is no longer an alternate UI in the works (as far as I understand it) and the production version has been rolled out everywhere in WW.