Update document to reflect confidence level change
There are few places where we are using 60, 70 & 80 as confidence levels which should be updated to 65, 75, 85. Also, we have certain screenshots in the document which are showing the older UI with confidence numbers while creating policies instead of the new UI with confidence levels. @chrfox - need your help in getting the images updated.