Update disposition.md
We currently don't log all the disposition actions into audit. But, we are working to get this done before we GA. I will confirm to you tomorrow on what exactly are logged into audit so that we can document it. I see the a couple of screenshots need to be replaced by newer ones which I am sending to you in Teams.