Update use-dmarc-to-validate-email.md
Removed instructions on adding DMARC record as it was pointing to a document which talks about adding MX and TXT records to connect your domain and nothing specific to DMARC. or another TXT record.

The document already contains instructions about forming DMARC records and best practices. So users will have to form the record and update it on their domain registrar for which the steps will be different depending on the host.