Clarification of role groups in ExO & Compliance
Role groups in Exchange Online and Microsoft Compliance are different. Membership in Exchange Online's Organization Management does not give permissions to perform a purge. 

I feel like this is not made clear in the current wording, and can be confusing to Exchange Administrators (including myself) encountering this how-to guide, thinking that they should have the appropriate permissions by already being a member of Organization Management in Exchange Online.

If one is only assigned Organization Management in Exchange Online rather than through Compliance, when importing the PowerShell cmdlets you will not be getting access to the -Purge parameter in New-ComplianceSearchAction, and it will instead throw an error.