Update cross-tenant-mailbox-migration.md
Troubleshooted migration attempt with customer today, and they ran into the issue that the move was failing because they only added the legacyExchangeDN to the target mailuser. But not the x500 addresses already present in the source mailbox. And MRS failed with more or less good error message. Therefore added this to the prereqs documentation.