Update multi-geo-capabilities-in-teams-in-microsoft-365.md
Got a little feedback from an MVP that we don't call out Private Channels specifically, so he wasn't sure if it would still follow the team or if it would follow the creator of the Private Channel. I think just adding this little bit will help clarify this question, but let me know if you think there is a better solution or place for the information.