Update register-devices-partner.md
Both the scripts are similar and it works for partners as well - So removed the part "he Windows PowerShell script is different from the one used for the Microsoft Managed Desktop Admin portal." from not. 

Also added a note about editing the file in excel and saving as a CSV will not generate a usable file -Extra columns are not supported. Quotes are not supported. Only ANSI-format text files can be used (not Unicode). Headers are case-sensitive. Editing the file in Excel and saving as a CSV file will not generate a usable file due to these requirements.