Update limits-for-content-search.md
Per support cases and an email request, modified and clarified that the max 1000 limit (row 6) for "items found in all user mailboxes that are displayed on the preview page when previewing the search results" to state that this is a theoretical maximum, and that the preview page is meant to show a sample of the results.

Also benign formatting fixes that makes it look like there were massive changes everywhere.