Update unlimited-archiving.md
This bug concerns this documentation: Overview of unlimited archiving - Microsoft 365 Compliance | Microsoft Docs

The documentation reads:
You can delete items in a subfolder that points to an auto-expanded storage area, but the folder itself can't be deleted.

However, this is incorrect. As long as the user has at least one AuxArchive (i.e. auto-expanded) location, no folder can be deleted from the AuxArchive or the MainArchive. The documentation should be updated to reflect this. It does not matter which content mailbox GUID the folder points to.