Update use-notifications-and-policy-tips.md
Added a sentence to the note in the section about policy tips in Outlook 2013 and later, based on a request from a CSS Support engineer that was verified and approved by Sr. SPM Gabe Bratton. @Chrfox, would you please review the new info, update for language as needed and merge the PR? Thank you!