Rename image files added in public to fix build warnings
This commit changes three file names (and the references to them) to resolve build warnings. The images were added in the public repo. The files are out of scope for the build, because "jpeg" is not an allowed type. I've renamed the files to use "jpg" instead, as well as adhere to guidance about file names (particularly, no capital letters).