Update permissions-in-the-security-and-compliance-center.md
Updating the article to remove references to the Attack Simulation training roles.  We had a recent scenario internally based on which some changes were made. Now, the role assignment for Attack Simulation administrator and attack payload author has to happen only via AAD and is unsupported/not possible through the defender portal.