Update teams-classes-with-canvas.md
Clarifying AAD and SDS requirements. Updating instructions to be in line with training materials and the modern onboarding LMS integration guide. Modifying headings/format updates to be in line with the other LTI integration instructions on docs.microsoft. Removing "Grant admin consent" section because the instructions are include in "Enable the Microsoft Teams App"