Update managing-office-365-endpoints.md
Proposed change is to call out in official documentation that not all published changes to the endpoints list will have Effective Dates lists. This is to clarify for customers and reduce confusion. Reference the following internal ICM for details from the team that owns the endpoints. https://portal.microsofticm.com/imp/v3/incidents/details/266294184/home