Diagnostic that can be run in context of M365 Administrator
I would like to provide this link as a button like what is seen in this article: https://docs.microsoft.com/en-us/sharepoint/troubleshoot/sharing-and-permissions/access-denied-due-to-org-policies 

This diagnostic is intended to identify site/page performance issues in the context as a Microsoft 365 Administrator. It will bring the user to the MAC portal and all they need to do is click Run.