Update office-365-air.md
Removing the separate alert policy table.  We're getting one table updated and the other gets missed on many of the recent changes.  Instead of duplicating a partial alert policy list, we will consolidate it with the default policy list - then add a column related to automated investigations over there.

Also made a slight change to the manually triggered investigations to reflect that these have alerts now.