Update web-content-filtering.md
Added a line stating that "Only Microsoft Edge is supported  and Network Protection is not supported on AVD Windows 10 multisession session hosts." As only Edge works on AVD Windows 10 multisession session hosts, as they are listed as OS configuration: Member Server. This might not be obvious to IT admins who may run into the issues.