Include missing word and delete extra words
Add missing word "be". Line 127 is now "In the list of people or role groups that can be added as members of the case...."
Delete extra words line 154: "the contains of" and "also". Sentence was "In addition to the exported data files, the contains of the export package also contains an export report, a summary report, and an error report." and is now " In addition to the exported data files, the export package contains an export report, a summary report, and an error report."