Important note on correct use of PowerShell Module
I have created a change where we need to be clearer on which PowerShell module we should use on this kind of checking. It is normal that users use the SCC powershell to see the logs where in the reality we must use the EXO. This was discussed in one ICM opened.