Update use-content-search-for-targeted-collections.md
based in my testing with customer and understanding of the get-mailboxfolderstatistics EXO cmdlet there is no limit of returned exchange folders. Therefore, suggesting to remove "The script displays a maximum of 100 mailbox folders." as this seems misleading.

Second change is regarding the amount of site folders of onedrive, in my tests and feedback from support + customer is, that only 100 site folders are returned not 200. Therefore, they also seems to be wrong.

Please check and approve to help customers save time with the right numbers here.