Update IRM Role Requirement Important Note
I work as a Support Engineer at Microsoft and I handled a case for this topic for a large customer of ours. They were not having proper alerts on their IRM portal. We finally got to the conclusiton that for IRM to work at least one admin needs to be a member of IRM roles, but although know that, we didn't know that the membership needs to be on the Built In IRM role. This is not clear in the Doc. My suggestion bellow:

>[!IMPORTANT]
>Make sure you always have at least one user in the Built-In *Insider Risk Management* or *Insider Risk Management Admin* role groups (depending on the option you choose) so that your insider risk management configuration doesn't get in to a 'zero administrator' scenario if specific users leave your organization.