Update defenderexpertsforhuntingprev.md
The most important part, the Apply link is missing. I've added it in. Please be more careful next time.

There are several other sections which are not updated based on what's in the public preview document. Those items are very important to set the expectations on feedback and such. Saving now. To be continued. I need to prep for dinner