Replaced slang terms
There is no such thing as the "Exchange Online admin" role.  There is no place in these docs for writers' comfort-zone slang.  Readers came here for a reason, we shouldn't steer them wrong with terminology that is not going to be found anywhere within the products that they've come here to learn about.