Update safe-links.md
First Change: 
Global Policy Settings menu and the policy configurations are in the process of being deprecated. Customers no longer have the ability to add new entries to the Block List in the Global Setting menu. Want to strike through the below text and make reference to TABL

"- **Global Safe Links settings**: These settings are configured globally, not in Safe Links policies. These settings include:

  - ["Block the following URLs" list for Safe Links](#block-the-following-urls-list-for-safe-links)"

Second Change: 
Removing the following example from the table as it is no longer possible due to change mentioned in number one.
"|In Lee's organization, `https://tailspintoys.com` is configured in the **Block the following URLs** list in the global settings for Safe Links. A Safe Links policy that includes Lee already exists. Lee receives an email message that contains the URL `https://tailspintoys.com/aboutus/trythispage`. Lee clicks the URL.|The URL might be automatically blocked for Lee; it depends on the URL entry in the list and the email client Lee used. For more information, see the ["Block the following URLs" list for Safe Links](#block-the-following-urls-list-for-safe-links) section later in this article.|"

Third Change: - Remove the following text "If the URL is included in the **Block the following URLs** list, a [blocked URL warning](#blocked-url-warning) opens."

Will remove the rest of the references to the Block URL List and Global Settings Menu when we are closer to deprecating it ETA: January 2023