M365 groups not supported for FromMemberOf
Line 71 has to be modified to match the Exchange DLP policy configurator from the Purview portal as M365 groups are not supported
In compliance.microsoft.com when you try to create an Exchange DLP policy with the Condition/Exception "sender is a member of" selected we see the following description:
Conditions
Define the conditions that must be met for this policy to be applied. Include specific content, senders, and recipients that you want the rule to detect. For more complex rules, create groups to exclude or include items. Learn how the condition builder works
Looks for messages sent from distribution groups, dynamic distribution groups or mail-enabled security groups.​
Upon selecting the option to add the distribution groups, only email enabled DLs, dynamic distribution groups created from Exchange admin portal and email-enabled security groups are available for selection