Update dlp-conditions-and-exceptions.md
Line 88 Recipient is member of needs to be changed to reflect what we have in the DLP policy configurator. As for the Sender is member of, M365 groups are not actually supported
What we see in the EXO DLP policy configurator is:

Looks for messages sent to members of distribution groups, dynamic distribution groups, or mail-enabled security groups.