Flow.microsoft.com is moving to make.powerautomate.com
Power automate is in the processing of moving its  portal to a new common Power Platform infrastructure which will include new domains for the Power Automate portal. Once we enable this, the existing Power Automate URLs will start redirecting to the new URLs. The old URL's ie. flow.microsoft.com will start redirecting to make.powerautomate.com. 

Hence the URL needs to be added to the Office 365 URLs and IP address ranges page.