Update ediscovery-content-search-reference.md
This entire paragraph needs to be removed. You CANNOT remove the license from a mailbox that is on hold. Doing so will introduce a validation error in Azure and the mailbox will fail to be disconnected/disabled due to the hold. This is NOT a supported state. Disconnected or de-licensed mailboxes are not searchable. Period. If content in a mailbox needs to be searchable, it must either be an active licensed user, a Shared mailbox (no license required) or be turned into an Inactive mailbox.