Update android-configure-mam.md
Line 34: Add a missing space between one sentence and the next.
Lines 50, 64, 76, 81, 86, 91, 111, and 114: Add a period to the sentence.
Lines 54, 100, 124, and 129: Delete extra spaces within the sentence or preceding it.
Lines 58, 62, 74, 84, 96, and 126: Indent the image to align with the text of the step. Add a period to the alt-text.
Lines 78-79, 82, 92, 94, and 100: Align the text with the text for the bullet point or number.
Lines 86-87, 91, 99, and 104: Delete HTML break code.
Line 105: Add a blank line to match the previous formatting.
Lines 81-108: This numbered list doesn't align with the bullet point it falls under, if indeed it is supposed to fall under that bulleted item, or follow the lettering convention of the previous steps under bullet points. The number of steps in this list doesn't match the images. The image at line 108 calls out a 4th step that doesn't exist here.  Line 88 says "Use this option to specify whether this policy applies to unmanaged devices," but this step doesn't include any options. Do you mean "Specify whether this policy applies to unmanaged devices"?
Line 109 and 127: Delete an extra blank line.
Line 112: Is Intune Company Portal the broker app? Why is this single item a bullet item? Would a sentence like this, "The broker app Intune Company Portal must be installed," make more sense? Or use active voice in a sentence like "Install the broker app Intune Company Portal, if you haven't already done so."
Lines 180-192 and 201-214: Renumbered to correct the incorrectly numbered items at lines 188 and 209.