Update custom-detection-rules.md
Change from: The columns NetworkMessageId and RecipientEmailAddress must be present to apply actions to email messages. 
Change to: The columns NetworkMessageId and RecipientEmailAddress must be in the present in the output results of the query to apply actions to email messages.

There is some confusion of the word 'present' where users think this just called in the KQL query, if they filter the output so that those column values are not visible then the results is that the action 'Move to mailbox folder' and 'Delete email' action is greyed out for the user and cannot be selected.