Update customer-key-availability-key-roll.md
We had a large customer 125k seats Rave case # 35048920 that initiated the roll process for the first time. They disabled the old key and saw clients unable to access mailboxes in growing numbers - re-enabled the disabled rolled key and were able to reconnect in 5 minutes for all mailboxes affected.  We only indicated not to delete this change changes that to delete or disable in the warning at the top. and also adds under Exchange DEP section "While this cmdlet starts the key roll operation for Exchange Online, the action doesn't complete immediately."