Update dlp-policy-reference.md
Could not find “Document name matches patterns” for Conditions SharePoint and Conditions OneDrive accounts in settings menu. It appears only for exchange 

made changes accordingly. If this is a bug or if the issue needs to be reported to PM, we can close the PR.

Fixes#https://github.com/MicrosoftDocs/microsoft-365-docs/issues/10980