Remove Additional resources webinar section
The mentioned webinar doesn't exist anymore, in fact none of the webinars exist anymore as apparently, they have been lost. Having this information in the article confuses our customers who start asking questions to us and ask for the links to these webinars. https://techcommunity.microsoft.com/t5/security-compliance-and-identity/using-sensitivity-labels-with-microsoft-teams-o365-groups-and/ba-p/1221885#M1380 if you click on the webinar link you are automatically redirected to Bing