Note about Override in Outlook in the Web
Hello Team,

While testing DLP policy with allow Override, I notice that for the end user in Outlook in the Web be able to see it (Override) within Policy Tip, the DLP policy needs to be "Turn On", while in "Test Mode (with notifications)" the Policy Tip appears but never with Override. The reason why this happens, is because DLP in test mode with notifications does not require for the user to do the override, because is not being yet blocked the "send" of the email containing sensitive information.

Plus is also necessary in the DLP configuration, Actions being configurated with some restrict access like "Block only people outside your organization".

All of this seems quite simple, however I could not find any Microsoft Documentation explaining these two points. Besides I am aware that are more colleagues in Microsoft struggling to see Override included in the Policy Tip (after selecting in the DLP configuration).

Therefore, if we could have some note, in documentation saying something like: "In Outlook in the Web to be able to see Override in Policy Tip, the DLP policy needs to be "Turn On", since in Test Mode (with notifications) it won´t be display. As well as, in DLP configuration some restriction action needs to be in place".