Remove 14 character requirement
The guidance in this doc does not align with our more updated recommendations for protecting M365 admins.  M365 admins should be cloud-only, password-less.  In addition, AAD only environments cannot enforce a 14-character password minimum.  I've consulted with the AAD PG on this and happy to discuss in detail as to why this recommendation is not valid.  Thanks!