Added note for parameters that trigger PSR trace
Currently we have S500 customers like siemens complaining that more parameters other than -c capture PSR traces which is not disclosed in the docs so when they collect logs with parameters:  -L, -H, -C, -I, -A, -V and -T they are also getting PSR traces captured an they are unaware of these. I have added a note to account for this so customers are informed.