Update cross-tenant-sharepoint-migration-step5.md
Updated the incompatibility note as follows to the correct period or time, and to match the Cross tenant documentation for OneDrive Migrations (https://learn.microsoft.com/en-us/microsoft-365/enterprise/cross-tenant-onedrive-migration-step5).
From:
"We recommend waiting a period of 24 hours. If your tenants are still reporting as incompatible, contact support."

To: 
"We recommend waiting a period of 48 hours. If your tenants are still reporting as incompatible, contact support."