Update loop-compliance-summary.md
removed * and ** with licensing information - we are not the experts on which licenses are required for each feature, and this information changes. we have linked to the documentation for each feature, we need to let that documentation speak to the licensing or we're going to be updating our documentation all the time and correcting silly errata.